Fix RGBcolor issue
Changed RGBcolor struct to short from int. I was getting an exception and this fixed it. i don't know why

+++ ConsoleApplication3/Constants.h
@@ -3,9 +3,9 @@ class Constants
 {
 public:
 	struct RGBcolor {
-		int red = 0;
-		int green = 0;
-		int blue = 0;
+		short red = 0;
+		short green = 0;
+		short blue = 0;
 	};
 
 	RGBcolor NONE;
